update tests for conference and list conference
