+++ package.json
@@ -23,7 +23,8 @@
     "eslint-plugin-jest": "^20.0.3",
     "flow-bin": "^0.49.1",
     "react-scripts": "1.0.7",
-    "react-test-renderer": "^15.6.1"
+    "react-test-renderer": "^15.6.1",
+    "redux-mock-store": "^1.2.3"
   },
   "scripts": {
     "predeploy": "npm run build",
