--- src/app/smoothie/conference/__tests__/ListConference.test.js
@@ -1,11 +1,0 @@
-// import React from 'react';
-// import { shallow } from 'enzyme';
-// import ListConference from '../ListConference';
-
-describe('conference component', () => {
-  it('should render conference component three times', () => {
-    // const listConference = shallow(<ListConference day={conferences[0]} />);
-    // expect(listConference.find('Conference').length).toBe(3);
-    expect(3).toBe(3);
-  });
-});
